Use PEP562 to delay import
