+++ numba/typed/__init__.py
@@ -1,2 +1,17 @@
-from .api import Dict
-from .api import List
+import importlib
+
+
+_delayed_symbols = {
+    "Dict": ".typeddict",
+    "List": ".typedlist",
+}
+
+
+def __getattr__(name):
+    # Uses PEP-562 but requires python>3.6
+    if name in _delayed_symbols:
+        modpath = _delayed_symbols[name]
+        mod = importlib.import_module(modpath, __name__)
+        return getattr(mod, name)
+    else:
+        return importlib.import_module(f".{name}", __name__)
